Begin adding Android support
