+++ hotham-cubeworld/Cargo.toml
@@ -11,5 +11,11 @@ version = "0.1.0"
 cgmath = "0.18"
 hotham = {path = "../hotham"}
 
+[target.'cfg(target_os = "android")'.dependencies]
+ndk-glue = "0.3.0"
+
+[target.'cfg(target_os = "android")'.lib]
+crate-type = ["lib", "cdylib"]
+
 [build-dependencies]
 shaderc = "0.7"
